[S163-KARG] Commit kernel arg patch + JSON guard + session docs
sieve_filter.py:
- [S163-KARG] cp.int32() explicit typing for n_seeds, k, skip_min, skip_max
  in run_sieve() non-hybrid path — matches existing hybrid path pattern
- Fixes OOB GPU memory access from int64/int32 kernel arg mismatch

window_optimizer_integration_final.py:
- JSON size guard: _JSON_WRITE_LIMIT=100K on all 3 write sites
- Prevents multi-hour json.dump hang when survivor accumulators exceed 100K records
- Full data preserved in NPZ files (downstream steps consume exclusively)

apply_s163_kernel_arg_typing.py:
- Patch script for sieve_filter.py kernel arg fix (dry-run + live apply)

docs/SESSION_CHANGELOG_20260418_S163KARG_FORENSIC.md:
- Full session record: crash analysis, forensic tooling upgrade, GDM finding

docs/TB_Incident_Report_rrig6600c_S163KARG.md:
- TB-reviewed incident report with corrected PERMISSION_FAULTS decode
- GDM/Xorg finding and TB verdict incorporated

+++ sieve_filter.py
@@ -256,10 +256,18 @@ class GPUSieve:
                 threads_per_block = 256
                 blocks = (n_seeds + threads_per_block - 1) // threads_per_block
                 # Build kernel arguments
+                # [S163-KARG] Explicit int32 typing — C++ kernel signature is int32.
+                # Bare Python ints marshal as int64 on x86_64 CuPy, shifting arg
+                # buffer layout by 4 bytes per arg. Misaligned args past this point
+                # read garbage (including n_seeds itself if misread earlier).
+                # Mirrors run_hybrid_sieve pattern (line ~435) which already uses
+                # explicit cp.int32() wrapping.
                 kernel_args = [
                     seeds_gpu, residues_gpu, survivors_gpu,
                     match_rates_gpu, best_skips_gpu, survivor_count_gpu,
-                    n_seeds, k, skip_min, skip_max, cp.float32(min_match_threshold)
+                    cp.int32(n_seeds), cp.int32(k),
+                    cp.int32(skip_min), cp.int32(skip_max),
+                    cp.float32(min_match_threshold),
                 ]
                 # Add PRNG-specific parameters
                 default_params = config.get("default_params", {})
